# docs/swagger.yaml
basePath: /api/v1
definitions:
  types.FailureErrorResponse:
    properties:
      error:
        type: string
      message:
        type: string
      status:
        type: string
    type: object
  types.FailureResponse:
    properties:
      message:
        type: string
      status:
        type: string
    type: object
...
  types.TeamCreateRequest:
    properties:
      name:
        type: string
    type: object
  types.TeamInviteRequest:
    properties:
      role:
        description: '"owner" или "member" (по умолчанию member)'
        type: string
      user_id:
        type: string
    type: object
  types.TeamListResponse:
    properties:
      teams:
        items:
          $ref: '#/definitions/types.TeamResponse'
        type: array
    type: object
  types.TeamResponse:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
info:
  contact: {}
  description: This is an API of auth-service
  title: Service task  Swagger
  version: "1.0"
paths:
...
  /teams:
    get:
      description: ListUserTeams возвращает список команд, где состоит текущий пользователь.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.TeamListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/types.FailureResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.FailureErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: listUserTeams info
      tags:
      - teams
    post:
      description: CreateTeam создает команду и делает текущего пользователя её owner.
      parameters:
      - description: team data
        in: body
        name: team
        required: true
        schema:
          $ref: '#/definitions/types.TeamCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.TeamResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/types.FailureResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.FailureErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: createTeam info
      tags:
      - teams
  /teams/{id}/invite:
    post:
      description: InviteToTeam приглашает пользователя в команду (может только owner).
      parameters:
      - description: team ID
        in: path
        name: id
        required: true
        type: string
      - description: invite data
        in: body
        name: invite
        required: true
        schema:
          $ref: '#/definitions/types.TeamInviteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.TeamListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/types.FailureResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.FailureErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: inviteToTeam info
      tags:
      - teams
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    description: Description for what is this security definition being used
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"